+++ src/commands/toggleConvertFileOnSave.ts
@@ -1,11 +1,23 @@
 import * as vscode from 'vscode';
-import registerConvertFileOnSave from '../registerConvertFileOnSave';
 
-const toggleConvertFileOnSave = vscode.commands.registerCommand('tkww.toggleConvertFileOnSave', () => {
-  const config = vscode.workspace.getConfiguration('TKWW');
-  const originalValue = config.get('convertFileOnSave');
-  config.update('convertFileOnSave', !originalValue);
-  registerConvertFileOnSave();
-});
+const createCommand = (outputChannel: vscode.OutputChannel) => {
+  const command = vscode.commands.registerCommand('tkww.toggleConvertFileOnSave', () => {
+    // Get configuration for TKWW
+    const config = vscode.workspace.getConfiguration('TKWW');
 
-export default toggleConvertFileOnSave;
+    // Get `convertFileOnSave` setting.
+    const originalValue = config.get('convertFileOnSave');
+
+    // Update the setting.
+    config.update('convertFileOnSave', !originalValue);
+
+    // Log to output channel
+    outputChannel.appendLine(`TKWW.convertFileOnSave was set to ${!originalValue}`);
+    vscode.window.showInformationMessage(`Files will ${!originalValue ? '' : 'not'} be modified on save.`);
+  });
+
+  return command;
+};
+
+
+export default createCommand;