+++ src/utils/index.ts
@@ -1,5 +1,7 @@
+export { default as accessors } from './accessors';
+export { default as constants } from './constants';
 export { default as findIndexFrom } from './findIndexFrom';
-export { default as getCase } from './getCase';
+export { default as getGitignoreGlob } from './getGitignoreGlob';
 export { default as moveCursor } from './moveCursor';
 export { default as startup } from './startup';
-export { default as updateText } from '../convertFile/updateText';
+export { default as updateText } from './updateText';